Pdf intialize

+++ SportsCashier/SportsCashier/Common/Models/SportsData.cs
@@ -11,13 +11,13 @@ namespace SportsCashier.Models
         public int Code { get; private set; }
         public string Name { get; private set; }
         public string Category { get; private set; }
-        public double Price { get; private set; }
+        public decimal Price { get; private set; }
         public bool CanDiscounted { get; private set; }
         public string Icon { get; private set; } = "\ue80e";
 
         public string NamePath => $"{Name} - {Category}";
 
-        public SportsData(int code, string name, string category, double price, bool canDiscounted, string icon)
+        public SportsData(int code, string name, string category, decimal price, bool canDiscounted, string icon)
         {
             Code = code;
             Name = name;
